server/pyproject.toml
[build-system]
requires = [ "poetry-core>=1.9.1",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "django-oscar-bluelight"
version = "5.6.0"
description = "Bluelight Specials - Enhancements to the offer and vouchers features for Django Oscar."
authors = ["thelab <thelabdev@thelab.co>"]
readme = "README.md"
homepage = "https://gitlab.com/thelabnyc/django-oscar-bluelight"
repository = "https://gitlab.com/thelabnyc/django-oscar-bluelight"
license = "ISC"

[[tool.poetry.packages]]
include = "oscarbluelight"
from = "src"

[tool.poetry.dependencies]
python = "^3.11"
celery = "^5.4.0"
Django = "^4.2.17"
django-oscar = "3.2.5"
thelabdb = "^0.1.2"
djangorestframework = "^3.15.2"

[tool.poetry.group.dev.dependencies]
coverage = "^7.6.10"
django-debug-toolbar = "^4.4.6"
django-redis = "^5.4.0"
flake8 = "^7.1.1"
hiredis = "^3.1.0"
psycopg2-binary = "^2.9.10"
redis = "^5.2.1"
sorl-thumbnail = "^12.11.0"
tox = "^4.23.2"
unittest-xml-reporting = "^3.2.0"
# Fix currently unpublished Python 3.12 compat
django-haystack = { git = "https://github.com/django-haystack/django-haystack.git", branch = "master" }
django-stubs = "^5.1.1"
mypy = "^1.14.1"
types-psycopg2 = "^2.9.21.20241019"
djangorestframework-stubs = "^3.15.2"
celery-types = "^0.22.0"

[tool.mypy]
python_version = "3.11"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

# Strict mode, see mypy --help
warn_unused_configs = true
# disallow_subclassing_any = true
# disallow_any_generics = true
# disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
# disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
# warn_return_any = true
no_implicit_reexport = true
show_error_codes = true
# Not turned on by strict
strict_equality = true

[[tool.mypy.overrides]]
module = "oscar.*"
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = "debug_toolbar.*"
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = "thelabdb.*"
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = "psycopg.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "*.tests.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "sandbox.settings"

[tool.isort]
profile = "black"
from_first = true

[tool.coverage.run]
branch = true
source_pkgs = ["oscarbluelight"]

[tool.coverage.report]
show_missing = true
ignore_errors = true
